requirements.txt
aiohug>=0.6
pyyaml
apispec==2.0.0
marshmallow==3.0.0rc4